pyproject.toml
[project]
name = "dpy-wav-sample"
version = "0.1.0"
description = "discord.pyでwavファイルをFFmpegなしで再生するサンプルプログラム。"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "audioop-lts>=0.2.1",
    "discord-py>=2.5.2",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.1",
]
